requirements.txt
httpx>=0.28.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=2.10.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
python-dotenv>=1.0.0
python-docx>=1.1.0
python-multipart>=0.0.9
# Auth
PyJWT>=2.8.0
# Rate limiting
slowapi>=0.1.9
# Billing
stripe>=7.0.0
# Twilio signature validation (optional)
twilio>=9.0.0
